estimate detailed results per variable and effect sizes for both fixed and random effects from lmer models

This function extends the current drop1 method for merMod class objects from the lme4 package. Where the default method to be able to drop both fixed and random effects at once.

# S3 method for merMod
modelTest(object, method = c("Wald", "profile", "boot"), control, ...)

Arguments

object
A merMod class object, the fitted result of lmer.

method
A character vector indicating the types of confidence intervals to calculate. One of “Wald”, “profile”, or “boot”.

control
A lmerControl() results used to control how models are estimated when updating.

...
Additional arguments passed to confint

Details

At the moment, the function is aimed to lmer models and has very few features for glmer or nlmer models. The primary motivation was to provide a way to provide an overall test of whether a variable “matters”. In multilevel data, a variable may be included in both the fixed and random effects. To provide an overall test of whether it matters requires jointly testing the fixed and random effects. This also is needed to provide an overall effect size.

The function works by generating a formula with one specific variable or “term” removed at all levels. A model is then fit on this reduced formula and compared to the full model passed in. This is a complex operation for mixed effects models for several reasons. Firstly, R has no default mechanism for dropping terms from both the fixed and random portions. Secondly, mixed effects models do not accomodate all types of models. For example, if a model includes only a random slope with no random intercept, if the random slope was dropped, there would be no more random effects, and at that point, lmer or glmer will not run the model. It is theoretically possible to instead fit the model using lm or glm but this becomes more complex for certain model comparisons and calculations and is not currently implemented. Marginal and conditional R2 values are calculated for each term, and these are used also to calculate something akin to an f-squared effect size.

This is a new function and it is important to carefully evaluate the results and check that they are accurate and that they are sensible. Check accuracy by viewing the model formulae for each reduced model and checking that those are indeed accurate. In terms of checking whether a result is sensible or not, there is a large literature on the difficulty interpretting main effect tests in the presence of interactions. As it is challenging to detect all interactions, especially ones that are made outside of R formulae, all terms are tested. However, it likely does not make sense to report results from dropping a main effect but keeping the interaction term, so present and interpret these with caution.

Examples

## these examples are slow to run
library(JWileymisc)
m1 <- lme4::lmer(extra ~ group + (1 | ID),
data = sleep, REML=FALSE)
modelTest(m1)

#> $FixedEffects
#>           Term  Est Type         LL       UL         Pval
#> 1: (Intercept) 0.75   FE -0.3663711 1.866371 1.879250e-01
#> 2:      group2 1.58   FE  0.8567759 2.303224 1.853388e-05
#> 
#> $RandomEffects
#>                 Term       Est Type LL UL
#> 1: sd_(Intercept)|ID 1.6010934   RE NA NA
#> 2:             sigma 0.8251061   RE NA NA
#> 
#> $EffectSizes
#>    Variable Estimator N_Obs N_Groups       AIC       BIC       LL LLDF
#> 1:    group               0          -8.414884 -7.419152 5.207442    1
#>         Sigma MarginalR2 ConditionalR2 MarginalF2 ConditionalF2     Chi2
#> 1: -0.5637784  0.1613328     0.3226657  0.1923682      1.833431 10.41488
#>              P   FE    RE              Formula  Type
#> 1: 0.001250037 TRUE FALSE extra ~ 1 + (1 | ID) Fixed
#> 
#> $OverallModel
#> $Performance
#>     Model Estimator N_Obs N_Groups      AIC      BIC        LL LLDF     Sigma
#> 1: merMod        ML    20  ID (10) 78.50469 82.48762 -35.25235    4 0.8251061
#>    MarginalR2 ConditionalR2 MarginalF2 ConditionalF2
#> 1:  0.1613328     0.8240099  0.1923682      4.682139
#> 
#> attr(,"class")
#> [1] "modelPerformance.merMod" "modelPerformance"       
#> 
#> attr(,"class")
#> [1] "modelTest.merMod" "modelTest"       


# \donttest{
data(aces_daily, package = "JWileymisc")

strictControl <- lme4::lmerControl(optCtrl = list(
   algorithm = "NLOPT_LN_NELDERMEAD",
   xtol_abs = 1e-10,
   ftol_abs = 1e-10))

m1 <- lme4::lmer(NegAff ~ STRESS + (1 + STRESS | UserID),
  data = aces_daily,
  control = strictControl)
modelTest(m1, method = "profile")

#> Computing profile confidence intervals ...

#> Parameters and CIs are based on REML, 
#> but modelTests requires ML not REML fit for comparisons, 
#> and these are used in effect sizes. Refitting.

#> $FixedEffects
#>           Term       Est Type        LL        UL Pval
#> 1: (Intercept) 1.2357011   FE 1.2024389 1.2692167    0
#> 2:      STRESS 0.1183128   FE 0.1068207 0.1296297    0
#> 
#> $RandomEffects
#>                             Term        Est Type         LL         UL
#> 1: cor_STRESS.(Intercept)|UserID 0.47080710   RE 0.28594503 0.63920496
#> 2:         sd_(Intercept)|UserID 0.20584754   RE 0.17824126 0.23575939
#> 3:              sd_STRESS|UserID 0.06769562   RE 0.05902194 0.07720673
#> 4:                         sigma 0.41789650   RE 0.41053223 0.42549600
#> 
#> $EffectSizes
#>    Variable Estimator N_Obs N_Groups        AIC        BIC        LL LLDF
#> 1:   STRESS               0          -2701.2231 -2680.9361 1353.6115    3
#> 2:   STRESS               0           -538.1871  -524.6625  271.0936    2
#>          Sigma  MarginalR2 ConditionalR2  MarginalF2 ConditionalF2      Chi2
#> 1: -0.10192042  0.22744094    0.11496663  0.29439943     0.2560813 2707.2231
#> 2: -0.02275066 -0.04844019    0.01044791 -0.06270095     0.0232721  542.1871
#>                P    FE   RE                            Formula           Type
#> 1:  0.000000e+00  TRUE TRUE          NegAff ~ 1 + (1 | UserID) Fixed + Random
#> 2: 1.843162e-118 FALSE TRUE NegAff ~ 1 + STRESS + (1 | UserID)         Random
#> 
#> $OverallModel
#> $Performance
#>     Model Estimator N_Obs     N_Groups      AIC      BIC        LL LLDF
#> 1: merMod        ML  6389 UserID (191) 7702.815 7743.389 -3845.407    6
#>        Sigma MarginalR2 ConditionalR2 MarginalF2 ConditionalF2
#> 1: 0.4179028  0.2274409     0.5510541  0.2943994       1.22744
#> 
#> attr(,"class")
#> [1] "modelPerformance.merMod" "modelPerformance"       
#> 
#> attr(,"class")
#> [1] "modelTest.merMod" "modelTest"       


m2 <- lme4::lmer(NegAff ~ STRESS + I(STRESS^2) + (1 + STRESS | UserID),
  data = aces_daily, control = strictControl)

## might normally use more bootstraps but keeping low for faster run
modelTest(m2, method = "boot", nsim = 100)

#> Computing bootstrap confidence intervals ...

#> 
#> 1 warning(s): Model failed to converge with max|grad| = 0.00222261 (tol = 0.002, component 1)

#> Parameters and CIs are based on REML, 
#> but modelTests requires ML not REML fit for comparisons, 
#> and these are used in effect sizes. Refitting.

#> $FixedEffects
#>           Term         Est Type          LL          UL         Pval
#> 1: (Intercept) 1.255682188   FE 1.218592743 1.287885765 0.000000e+00
#> 2: I(STRESS^2) 0.003907478   FE 0.002066481 0.005704836 4.873061e-05
#> 3:      STRESS 0.092421665   FE 0.075694182 0.109085934 0.000000e+00
#> 
#> $RandomEffects
#>                             Term        Est Type         LL        UL
#> 1: cor_STRESS.(Intercept)|UserID 0.46480701   RE 0.26388784 0.6916568
#> 2:         sd_(Intercept)|UserID 0.21407800   RE 0.18924474 0.2440853
#> 3:              sd_STRESS|UserID 0.06437972   RE 0.05339639 0.0701088
#> 4:                         sigma 0.41744838   RE 0.41013609 0.4266621
#> 
#> $EffectSizes
#>       Variable Estimator N_Obs N_Groups        AIC         BIC         LL LLDF
#> 1:      STRESS               0          -520.66418 -500.377178 263.332089    3
#> 2:      STRESS               0          -438.87694 -425.352274 221.438470    2
#> 3: I(STRESS^2)               0           -13.62524   -6.862912   7.812622    1
#>            Sigma  MarginalR2 ConditionalR2  MarginalF2 ConditionalF2      Chi2
#> 1: -0.0218673846 -0.02493856   0.013746427 -0.03270762   0.030804834 526.66418
#> 2: -0.0191993031 -0.04324950   0.007824503 -0.05672292   0.017534193 442.87694
#> 3: -0.0004769974  0.01008702   0.002707476  0.01322941   0.006067274  15.62524
#>                P    FE    RE                                          Formula
#> 1: 7.940652e-114  TRUE  TRUE          NegAff ~ 1 + I(STRESS^2) + (1 | UserID)
#> 2:  6.768529e-97 FALSE  TRUE NegAff ~ 1 + STRESS + I(STRESS^2) + (1 | UserID)
#> 3:  7.721681e-05  TRUE FALSE      NegAff ~ 1 + STRESS + (1 + STRESS | UserID)
#>              Type
#> 1: Fixed + Random
#> 2:         Random
#> 3:          Fixed
#> 
#> $OverallModel
#> $Performance
#>     Model Estimator N_Obs     N_Groups     AIC      BIC        LL LLDF    Sigma
#> 1: merMod        ML  6389 UserID (191) 7689.19 7736.526 -3837.595    7 0.417426
#>    MarginalR2 ConditionalR2 MarginalF2 ConditionalF2
#> 1:  0.2375305     0.5537575  0.3115279      1.240934
#> 
#> attr(,"class")
#> [1] "modelPerformance.merMod" "modelPerformance"       
#> 
#> attr(,"class")
#> [1] "modelTest.merMod" "modelTest"       

# }